new interface for generating parts of the information for a social card
